Feature specs to have a Nomis connection
