+++ spec/fixtures/vcr_cassettes/lookup_active_offender-error.yml
@@ -1,0 +1,30 @@
+---
+http_interactions:
+- request:
+    method: get
+    uri: http://172.22.16.2:8080/nomisapi/lookup/active_offender?date_of_birth=1976-06-12&noms_id=A1459AE
+    body:
+      encoding: US-ASCII
+      string: ''
+    headers:
+      User-Agent:
+      - excon/0.47.0
+      Accept:
+      - application/json
+  response:
+    status:
+      code: 500
+      message: ''
+    headers:
+      Server:
+      - Apache-Coyote/1.1
+      Content-Type:
+      - application/json
+      Date:
+      - Tue, 29 Mar 2016 14:49:29 GMT
+    body:
+      encoding: UTF-8
+      string: 'internal server error'
+    http_version:
+  recorded_at: Tue, 29 Mar 2016 14:49:54 GMT
+recorded_with: VCR 3.0.1
