+++ spec/fixtures/vcr_cassettes/process_booking_happy_path-dup.yml
@@ -1,0 +1,168 @@
+---
+http_interactions:
+- request:
+    method: get
+    uri: http://172.22.16.2:8080/nomisapi/lookup/active_offender?date_of_birth=1976-06-12&noms_id=A1459AE
+    body:
+      encoding: US-ASCII
+      string: ''
+    headers:
+      User-Agent:
+      - excon/0.49.0
+      Accept:
+      - application/json
+      Authorization:
+      - ''
+      X-Request-Id:
+      - 37944dba-eebb-4bcf-a5f4-5f5bfe3abe55
+  response:
+    status:
+      code: 200
+      message: ''
+    headers:
+      Server:
+      - Apache-Coyote/1.1
+      Server-Timing:
+      - total=95, db=77
+      Content-Type:
+      - application/json
+      Date:
+      - Thu, 01 Dec 2016 13:52:10 GMT
+    body:
+      encoding: UTF-8
+      string: '{"found":true,"offender":{"id":1055827}}'
+    http_version: 
+  recorded_at: Thu, 01 Dec 2016 13:52:14 GMT
+- request:
+    method: get
+    uri: http://172.22.16.2:8080/nomisapi/offenders/1055827/visiting_availability?end_date=2016-12-12&offender_id=1055827&start_date=2016-12-12
+    body:
+      encoding: US-ASCII
+      string: ''
+    headers:
+      User-Agent:
+      - excon/0.49.0
+      Accept:
+      - application/json
+      Authorization:
+      - ''
+      X-Request-Id:
+      - 37944dba-eebb-4bcf-a5f4-5f5bfe3abe55
+  response:
+    status:
+      code: 200
+      message: ''
+    headers:
+      Server:
+      - Apache-Coyote/1.1
+      Server-Timing:
+      - total=442, db=439
+      Content-Type:
+      - application/json
+      Date:
+      - Thu, 01 Dec 2016 13:52:10 GMT
+    body:
+      encoding: UTF-8
+      string: '{"available":true,"dates":["2016-12-12"]}'
+    http_version: 
+  recorded_at: Thu, 01 Dec 2016 13:52:15 GMT
+- request:
+    method: get
+    uri: http://172.22.16.2:8080/nomisapi/lookup/active_offender?date_of_birth=1976-06-12&noms_id=A1459AE
+    body:
+      encoding: US-ASCII
+      string: ''
+    headers:
+      User-Agent:
+      - excon/0.49.0
+      Accept:
+      - application/json
+      Authorization:
+      - ''
+      X-Request-Id:
+      - d04d22a9-7475-403c-8657-b6dd50f62294
+  response:
+    status:
+      code: 200
+      message: ''
+    headers:
+      Server:
+      - Apache-Coyote/1.1
+      Server-Timing:
+      - total=3, db=1
+      Content-Type:
+      - application/json
+      Date:
+      - Thu, 01 Dec 2016 13:52:11 GMT
+    body:
+      encoding: UTF-8
+      string: '{"found":true,"offender":{"id":1055827}}'
+    http_version: 
+  recorded_at: Thu, 01 Dec 2016 13:52:15 GMT
+- request:
+    method: get
+    uri: http://172.22.16.2:8080/nomisapi/offenders/1055827/visiting_availability?end_date=2016-12-12&offender_id=1055827&start_date=2016-12-12
+    body:
+      encoding: US-ASCII
+      string: ''
+    headers:
+      User-Agent:
+      - excon/0.49.0
+      Accept:
+      - application/json
+      Authorization:
+      - ''
+      X-Request-Id:
+      - d04d22a9-7475-403c-8657-b6dd50f62294
+  response:
+    status:
+      code: 200
+      message: ''
+    headers:
+      Server:
+      - Apache-Coyote/1.1
+      Server-Timing:
+      - total=4, db=2
+      Content-Type:
+      - application/json
+      Date:
+      - Thu, 01 Dec 2016 13:52:11 GMT
+    body:
+      encoding: UTF-8
+      string: '{"available":true,"dates":["2016-12-12"]}'
+    http_version: 
+  recorded_at: Thu, 01 Dec 2016 13:52:15 GMT
+- request:
+    method: get
+    uri: http://172.22.16.2:8080/nomisapi/lookup/active_offender?date_of_birth=1976-06-12&noms_id=A1459AE
+    body:
+      encoding: US-ASCII
+      string: ''
+    headers:
+      User-Agent:
+      - excon/0.49.0
+      Accept:
+      - application/json
+      Authorization:
+      - ''
+      X-Request-Id:
+      - 55040ee6-7ab9-42a5-9e4c-175a2a96ea55
+  response:
+    status:
+      code: 200
+      message: ''
+    headers:
+      Server:
+      - Apache-Coyote/1.1
+      Server-Timing:
+      - total=3, db=1
+      Content-Type:
+      - application/json
+      Date:
+      - Thu, 01 Dec 2016 13:52:11 GMT
+    body:
+      encoding: UTF-8
+      string: '{"found":true,"offender":{"id":1055827}}'
+    http_version: 
+  recorded_at: Thu, 01 Dec 2016 13:52:16 GMT
+recorded_with: VCR 3.0.1
